[Feat] Search to Music Intent Logic Add
