+++ feature/home/src/main/java/com/mument_android/home/util/ExtBindingAdapter.kt
@@ -7,17 +7,3 @@ import com.mument_android.domain.entity.home.RecentSearchData
 import com.mument_android.home.adapters.SearchListAdapter
 
 
-@BindingAdapter("app:ui_state_search_result")
-fun RecyclerView.bindUiStateSearchResultList(uiState: UiState){
-    val boundAdapter = this.adapter
-    visibility = if (boundAdapter is SearchListAdapter && uiState is UiState.Success<*>) {
-        (uiState.data as List<RecentSearchData>).run {
-            boundAdapter.submitList(this)
-        }
-        View.VISIBLE
-    } else {
-        View.GONE
-    }
-}
-
-
